server/requirements.txt
fastapi
uvicorn
sqlalchemy
openai
pydantic-settings
python-dotenv
httpx
